+++ src/Lit/Air/Tests/AutowireRecipeTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Lit\Air\Tests;
 
 use Lit\Air\Injection\SetterInjector;
-use Lit\Air\Psr\Container;
+use Lit\Air\Recipe\AbstractRecipe;
 use Lit\Air\Recipe\AutowireRecipe;
 
 class AutowireRecipeTest extends AirTestCase
@@ -14,7 +14,7 @@ class AutowireRecipeTest extends AirTestCase
     {
         $key = self::randKey();
         $obj = new \stdClass();
-        $recipe = Container::autowire(Foo::class, [
+        $recipe = AbstractRecipe::autowire(Foo::class, [
             'bar' => $obj,
         ]);
         $this->assertInstanceOf(AutowireRecipe::class, $recipe);
@@ -31,7 +31,7 @@ class AutowireRecipeTest extends AirTestCase
 
         $key2 = self::randKey();
         $this->getFactory()->addInjector(new SetterInjector());
-        $this->container->define($key2, Container::instance(Foo::class, [
+        $this->container->define($key2, AbstractRecipe::instance(Foo::class, [
             'bar' => $obj,
         ]));
         $instance = $this->container->get($key2);
