+++ src/Lit/Air/Tests/FixedValueRecipeTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Lit\Air\Tests;
 
-use Lit\Air\Psr\Container;
+use Lit\Air\Recipe\AbstractRecipe;
 use Lit\Air\Recipe\FixedValueRecipe;
 
 class FixedValueRecipeTest extends AirTestCase
@@ -13,7 +13,7 @@ class FixedValueRecipeTest extends AirTestCase
     {
         $key = self::randKey();
         $obj = new \stdClass();
-        $stub = Container::value($obj);
+        $stub = AbstractRecipe::value($obj);
         self::assertTrue($stub instanceof FixedValueRecipe);
 
         $returnValue = $this->container->define($key, $stub);
